scripts/requirements.txt
# CangjieCorpus → vector DB pipeline
chromadb>=0.4.22
sentence-transformers>=2.2.2
GitPython>=3.1.40
